Extract rollups + step-status derivation from workspace_status.py
Splits modules/workspace_status.py (879 lines) into three thematic
files, keeping the top-level orchestrators in place.

## What moved

### modules/workspace_status_constants.py (~76 lines, new)

Bottom-layer file holding the QS_* constants:
- QS_REQUIRED_STEP_KEYS, QS_REVIEW_STEP_KEYS, QS_VALIDATION_STEP_KEYS
- QS_STATUS_ORDER, QS_WARN_REASONS, QS_ERROR_REASONS
- QS_FINAL_VALIDATION_TTL_HOURS

Nothing in this file imports from any sibling.  Breaks the
circular-import problem cleanly.

### modules/workspace_rollups.py (~300 lines, new)

11 helper functions grouped as:
- Status rollups: _worst_status, _derive_live_final_validation_status,
  _build_live_validation_rollup
- Timestamp helpers: _latest_iso_timestamp, _format_validation_age,
  _parse_iso_datetime, _bulk_validation_is_fresh
- Final gate + navigation: _build_final_gate, _step_href,
  _latest_bulk_validation_timestamp, _workspace_step_status_from_app_readiness

### modules/workspace_step_status.py (~252 lines, new)

4 functions for per-step status derivation:
- _is_nonblank_setting, _is_meaningful_optional_status_input
- _has_meaningful_optional_input, _derive_step_status

## modules/workspace_status.py: 879 -> 414 (-465 / -52.9%)

The 414-line file now contains only docstring + 4 re-export blocks +
the three orchestrator functions (_build_workspace_status_context and
the two _build_workspace_app_readiness variants).

## Backward compatibility

workspace_status.py re-exports every moved name.  All 26 imports in
quickstart.py continue to work unchanged.  No test-side changes
needed.

## Verification

AST-verified: all 15 moved functions are byte-for-byte identical
to origin/develop.  Full test suite passes.  Ruff + Black clean.
